+++ rmf_task/include/rmf_task/requests/ChargeBattery.hpp
@@ -53,6 +53,9 @@ public:
     /// Generate the description for this request
     static Task::ConstDescriptionPtr make();
 
+    /// Make a charging task that will last indefinitely.
+    static std::shared_ptr<Description> make_indefinite();
+
     // Documentation inherited
     Task::ConstModelPtr make_model(
       rmf_traffic::Time earliest_start_time,
@@ -62,6 +65,16 @@ public:
     Info generate_info(
       const State& initial_state,
       const Parameters& parameters) const final;
+
+    /// Set the charging task to run indefinitely. This means it will never
+    /// declare itself as finished and must instead be canceled. This can be
+    /// used for idle tasks that are canceled automatically when a task request
+    /// comes in. If indefinite is false, the robot will charge up to its
+    /// designated recharge level.
+    void set_indefinite(bool value);
+
+    /// Should this recharge task run indefinitely?
+    bool indefinite() const;
 
     class Implementation;
   private: