+++ rmf_task/src/rmf_task/requests/ChargeBatteryFactory.cpp
@@ -17,9 +17,31 @@
 
 #include <rmf_task/requests/ChargeBatteryFactory.hpp>
 #include <rmf_task/requests/ChargeBattery.hpp>
+#include <random>
 
 namespace rmf_task {
 namespace requests {
+
+//==============================================================================
+namespace {
+std::string generate_uuid(const std::size_t length = 3)
+{
+  std::stringstream ss;
+  for (std::size_t i = 0; i < length; ++i)
+  {
+    std::random_device rd;
+    std::mt19937 gen(rd());
+    std::uniform_int_distribution<> dis(0, 255);
+    const auto random_char = dis(gen);
+    std::stringstream hexstream;
+    hexstream << std::hex << random_char;
+    auto hex = hexstream.str();
+    ss << (hex.length() < 2 ? '0' + hex : hex);
+  }
+  return ss.str();
+}
+
+} // anonymous namespace
 
 //==============================================================================
 class ChargeBatteryFactory::Implementation
@@ -27,6 +49,7 @@ class ChargeBatteryFactory::Implementation
 public:
   std::optional<std::string> requester;
   std::function<rmf_traffic::Time()> time_now_cb;
+  bool indefinite = false;
 };
 
 //==============================================================================
@@ -48,22 +71,47 @@ ChargeBatteryFactory::ChargeBatteryFactory(
 }
 
 //==============================================================================
+void ChargeBatteryFactory::set_indefinite(bool value)
+{
+  _pimpl->indefinite = value;
+}
+
+//==============================================================================
+bool ChargeBatteryFactory::indefinite() const
+{
+  return _pimpl->indefinite;
+}
+
+//==============================================================================
 ConstRequestPtr ChargeBatteryFactory::make_request(const State& state) const
 {
-
+  const std::string id = "Charge" + generate_uuid();
+  Task::ConstBookingPtr booking;
   if (_pimpl->requester.has_value() && _pimpl->time_now_cb)
   {
-    return ChargeBattery::make(
+    booking =
+      std::make_shared<const rmf_task::Task::Booking>(
+      std::move(id),
       state.time().value(),
+      nullptr,
       _pimpl->requester.value(),
       _pimpl->time_now_cb(),
+      true);
+  }
+  else
+  {
+    booking =
+      std::make_shared<const rmf_task::Task::Booking>(
+      std::move(id),
+      state.time().value(),
       nullptr,
       true);
   }
-  return ChargeBattery::make(
-    state.time().value(),
-    nullptr,
-    true);
+
+  const auto description = ChargeBattery::Description::make_indefinite();
+  description->set_indefinite(_pimpl->indefinite);
+
+  return std::make_shared<Request>(std::move(booking), std::move(description));
 }
 
 } // namespace requests